refactor: 소스코드 refactor + test를 위한 코드 추가
